+++ ext/duckdb/appender.c
@@ -93,14 +93,27 @@ static VALUE appender_initialize(VALUE self, VALUE con, VALUE schema, VALUE tabl
  */
 static VALUE appender_error_message(VALUE self) {
     rubyDuckDBAppender *ctx;
-    const char *msg;
+#ifdef HAVE_DUCKDB_H_GE_V1_4_0
+    duckdb_error_data error_data;
+#endif
+    const char *msg = NULL;
+    VALUE rb_msg = Qnil;
     TypedData_Get_Struct(self, rubyDuckDBAppender, &appender_data_type, ctx);
 
-    msg = duckdb_appender_error(ctx->appender);
-    if (msg == NULL) {
-        return Qnil;
+#ifdef HAVE_DUCKDB_H_GE_V1_4_0
+    error_data = duckdb_appender_error_data(ctx->appender);
+    if (duckdb_error_data_has_error(error_data)) {
+        msg = duckdb_error_data_message(error_data);
+        rb_msg = rb_str_new2(msg);
     }
-    return rb_str_new2(msg);
+    duckdb_destroy_error_data(&error_data);
+#else
+    msg = duckdb_appender_error(ctx->appender);
+    if (msg != NULL) {
+        rb_msg = rb_str_new2(msg);
+    }
+#endif
+    return rb_msg;
 }
 
 /* :nodoc: */