+++ ext/duckdb/logical_type.h
@@ -9,5 +9,5 @@ typedef struct _rubyDuckDBLogicalType rubyDuckDBLogicalType;
 
 void rbduckdb_init_duckdb_logical_type(void);
 VALUE rbduckdb_create_logical_type(duckdb_logical_type logical_type);
-
+rubyDuckDBLogicalType *get_struct_logical_type(VALUE obj);
 #endif