+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
+# Unreleased
+- Support TIMESTAMP_NS infinity, -infinity value.
+
 # 1.3.1.0 - 2025-06-28
 - Support TIMESTAMP_S, TIMESTAMP_MS infinity, -infinity value.
 - bump duckdb to 1.3.1 on CI.