+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 # Unreleased
+
+# 1.4.1.0 - 2025-11-01
 - add DuckDB::Connection#appender_from_query.
 - add DuckDB::Appender.create_query.
 - add `DuckDB::LogicalType.boolean`, `DuckDB::LogicalType.tinyint`, `DuckDB::LogicalType.smallint`,